+++ tests/NPipeline.Extensions.Observability.Tests/LoggingPipelineMetricsSinkTests.cs
@@ -22,7 +22,7 @@ public sealed class LoggingPipelineMetricsSinkTests
         return logger;
     }
 
-    private static IPipelineMetrics CreatePipelineMetrics(bool success, Exception? exception = null, IReadOnlyList<INodeMetrics>? nodeMetrics = null, long totalItemsProcessed = 285, long? durationMs = 5000)
+    private static IPipelineMetrics CreatePipelineMetrics(bool success, Exception? exception = null, IReadOnlyList<INodeMetrics>? nodeMetrics = null, long totalItemsProcessed = 285, double? durationMs = 5000)
     {
         return new PipelineMetrics("TestPipeline", s_pipelineId, Guid.NewGuid(), DateTimeOffset.UtcNow.AddSeconds(-5), DateTimeOffset.UtcNow, durationMs, success, totalItemsProcessed, nodeMetrics ?? [], exception);
     }
